Call the llvm intrinsic sqrt function instead of an extern function
LLVM has an intrinsic sqrt function that translates to a single assembly
instruction for most targets. Instead of calling an extern function call
the intrinsic instruction directly.

I did this by replacing the extern declaration of 'proc sqrt' with an inline
version that calls a new primitive. I threaded the new primitive through the
various switches over every primitive and added a new codegen function to
generate the intrinsic call.

I added a test that can be compiled with '--savec output' to see that it
is now calling the the 'sqrt' assembly instruction directly instead of
through another function.

For ptx, this generates a direct use of the 'sqrt.rn.f64' instruction
instead of a call to '__nv_sqrt'.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>